attempt at fix#2

+++ src/phev_connect.js
@@ -21,58 +21,60 @@ const mqttUri = process.env.MQTT_URI || 'mqtt://secure.wattu.com'
 const PhevConnect = ({ mqtt } = {}) => {
 
     const phevMqtt = PhevMqtt({ mqtt, uri: mqttUri, options: { username: mqttUsername, password: mqttPassword } })
-    
+
     const client = new net.Socket();
-    
+
     const connectToCar = () => {
-        
+
         log.info('Connecting to ' + carHost + ':' + carPort);
-        client.connect(carPort, carHost, () => {
-            
-        })  
-        
-        client.on('connect', () => {
-            log.debug('Client connected')
-            phevMqtt.send('connection', 'connected')
-            Observable.fromEvent(client, 'data').subscribe(data => {
-                phevMqtt.send(phevReceive, data)
-                log.debug('Car    : ' + data.toString('hex'))
+
+        if (!client.writable) {
+            client.connect(carPort, carHost, () => {
+
             })
-        })
-        client.on('end', () => {
-            log.debug('Client socket ended')
-            client.destroy()
-        })
-        client.on('error', err => {
-            log.error('Client socket error ' + err)
-            phevMqtt.send(phevError, err)
-            client.destroy()
-        })      
+
+            client.on('connect', () => {
+                log.debug('Client connected')
+                phevMqtt.send('connection', 'connected')
+                Observable.fromEvent(client, 'data').subscribe(data => {
+                    phevMqtt.send(phevReceive, data)
+                    log.debug('Car    : ' + data.toString('hex'))
+                })
+            })
+            client.on('end', () => {
+                log.debug('Client socket ended')
+                client.destroy()
+            })
+            client.on('error', err => {
+                log.error('Client socket error ' + err)
+                phevMqtt.send(phevError, err)
+                client.destroy()
+            })
+        }
+
+        const connectToMqtt = () => {
+
+            log.debug('Subscribed ' + mqttUri)
+            phevMqtt.subscribe(phevSend)
+
+            phevMqtt.messages(phevSend)
+                .subscribe(m => {
+
+                    if (!client.destroyed) {
+                        client.write(m.message)
+                    } else {
+                        connectToCar()
+                        client.once('connect', () => client.write(m.message))
+                    }
+                })
+        }
+        const connect = () => {
+            connectToMqtt()
+        }
+
+        return {
+            connect: connect
+        }
     }
 
-    const connectToMqtt = () => {
-        
-        log.debug('Subscribed ' + mqttUri)
-        phevMqtt.subscribe(phevSend)
-
-        phevMqtt.messages(phevSend)
-            .subscribe(m => {
-
-            if(!client.destroyed) {
-                client.write(m.message)
-            } else {
-                connectToCar()
-                client.once('connect', () => client.write(m.message))
-            }
-        })
-    }
-    const connect = () => {
-        connectToMqtt()
-    }
-
-    return {
-        connect: connect
-    }
-}
-
-export default PhevConnect
+    export default PhevConnect
